+++ tests/test_make.py
@@ -305,10 +305,13 @@ class TestAttributes(object):
         assert C.D.__name__ == "D"
         assert C.D.__qualname__ == C.__qualname__ + ".D"
 
-    def test_post_init(self):
+    @given(with_validation=booleans())
+    def test_post_init(self, with_validation, monkeypatch):
         """
         Verify that __attrs_post_init__ gets called if defined.
         """
+        monkeypatch.setattr(_config, '_run_validators', with_validation)
+
         @attributes
         class C(object):
             x = attr()