+++ CHANGELOG.rst
@@ -24,8 +24,8 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-- Raise ``FrozenInstanceError`` when trying to delete an attribute from
-  a frozen class. `#118 <https://github.com/hynek/attrs/pull/118>`__
+- Raise ``FrozenInstanceError`` when trying to delete an attribute from a frozen class.
+  `#118 <https://github.com/hynek/attrs/pull/118>`_
 
 ----